Refactor to give many callset-specific outputs
Instead of each component having one output with multiple fields,
this goes back to giving each component many outputs.

Further, each callset can specify a subset of outputs that it triggers,
hopefully allowing for some replication of the generic written-tracking
that the more general circuit gave.

Another model is that this is a lot more like many distinct components
operating on the same shared state, with the restriction that any
given component can only be called in one trigger